+++ crates/core/src/flow/resolve.rs
@@ -48,7 +48,9 @@ impl Tool {
                     )
                     .await?;
 
-                self.inventory.save_remote_versions(&versions)?;
+                if !versions.versions.is_empty() {
+                    self.inventory.save_remote_versions(&versions)?;
+                }
             }
         }
 
@@ -243,7 +245,7 @@ impl Tool {
 
         let resolver = self.load_version_resolver(&candidate).await?;
         let version = self
-            .resolve_version_candidate(&resolver, &candidate, true)
+            .resolve_version_candidate(&resolver, &candidate, spec.resolve_from_manifest)
             .await?;
 
         debug!(